Configured the system to try and run make on main.cpp
Currently we are unable to build due to missing the 'uint' definition. We can define these in a 'types.h' header.

Using CLion IDE -- this system/project was configured using MinGW (installed packages gcc, g++, mingw32-make, mingw32-zlib), the CLion built-in version of CMake (v3.3) and the CLion built-in version of GDB.